ftdi-hal: access to gpio pins
The first attempt to implement 'controlled' access to gpio lines
available on FT232H chip. The idea is to be able to 'instantiate'
each pin only once.

This time implementation is based on RefCell<bool> fields which
are initialized as true, but reset to false when pin is instantiated

NB: this is not needed for spi/i2c protocol buses since generally
multiple devices can be connected to the same i2c/spi bus.

However we may need to impose the followin constraint: either
i2c or spi bus can be instantiated. This probably should be
done since h/w circuitry differs between spi and i2c.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>