examples: add support for both ftdi-rs and libftd2xx-rs
Support both ftdi-rs and libftd2xx-rs libraries in examples. For this
purpose use features 'ftdi-lib' and 'ftd2-lib' until a better option
is implemented. Note that these features are relevant only for
examples, not for the crate itself. It is ensured that exactly
one of those features is selected.

Example build commands for ftdi-rs backend:

$ cargo test --features=ftdi-lib --example=at24c04-test1 -- --test-threads=1
$ cargo run --features=ftdi-lib --example mcp3008-test1
$ cargo run --features=ftdi-lib --example ws2812-test1
$ cargo build --features=ftdi-lib --example lm75-test

Example build commands for libftd2xx-rs backend:

$ cargo test --features=ftd2-lib --example=at24c04-test1 -- --test-threads=1 test::at24x_test_t3
$ cargo run --features=ftd2-lib --example mcp3008-test1
$ cargo run --features=ftd2-lib --example ws2812-test1
$ cargo build --features=ftd2-lib --example lm75-test1

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>